websocket: split endpoints for replication, agents, publishers and listeners
